+++ chopper/lib/src/browser.dart
@@ -1,0 +1,4 @@
+import 'package:http/http.dart';
+import 'package:http/browser_client.dart';
+
+Client createHttpClient() => BrowserClient();